requirements.txt
pre-commit
python-dotenv
Pillow
imageHash